chatbot/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
pypdf
psycopg2-binary